+++ packages/editor/packages/editor-state/src/effects/codeBlocks/codeBlockDecorators/inputs/updateGraphicData.ts
@@ -2,7 +2,7 @@ import { getModuleId } from '@8f4e/compiler/syntax';
 
 import parseInputs from './codeParser';
 
-import gapCalculator from '../../../../pureHelpers/codeEditing/gapCalculator';
+import gapCalculator from '../../../../features/code-editing/gapCalculator';
 
 import type { CodeBlockGraphicData, State } from '../../../../types';
