+++ packages/editor/packages/editor-state/src/effects/codeBlocks/codeBlockDecorators/outputs/updateGraphicData.ts
@@ -2,7 +2,7 @@ import { getModuleId } from '@8f4e/compiler/syntax';
 
 import parseOutputs from './codeParser';
 
-import gapCalculator from '../../../../pureHelpers/codeEditing/gapCalculator';
+import gapCalculator from '../../../../features/code-editing/gapCalculator';
 
 import type { CodeBlockGraphicData, Output, State } from '../../../../types';
