+++ packages/editor/packages/editor-state/src/effects/codeBlocks/codeBlockDecorators/switches/updateGraphicData.ts
@@ -1,6 +1,6 @@
 import parseSwitches from './codeParser';
 
-import gapCalculator from '../../../../pureHelpers/codeEditing/gapCalculator';
+import gapCalculator from '../../../../features/code-editing/gapCalculator';
 
 import type { CodeBlockGraphicData, State } from '../../../../types';
