+++ packages/editor/packages/editor-state/src/effects/codeBlocks/creationIndex.test.ts
@@ -4,7 +4,7 @@ import createStateManager from '@8f4e/state-manager';
 import codeBlockCreator from './codeBlockCreator';
 import graphicHelper from './graphicHelper';
 
-import { flattenProjectForCompiler } from '../../features/program-compiler/compiler';
+import { flattenProjectForCompiler } from '../../features/program-compiler/effect';
 import projectImport from '../projectImport';
 import { createMockState, createMockCodeBlock } from '../../pureHelpers/testingUtils/testUtils';
 import { createMockEventDispatcherWithVitest } from '../../pureHelpers/testingUtils/vitestTestUtils';
