+++ packages/editor/packages/editor-state/src/features/code-editing/backSpace.ts
