+++ packages/editor/packages/editor-state/src/features/code-editing/gapCalculator.ts
