+++ packages/editor/packages/editor-state/src/features/code-editing/highlightSyntax8f4e.ts
