+++ packages/editor/packages/editor-state/src/features/program-compiler/disableCompilation.test.ts
@@ -1,9 +1,9 @@
 import { describe, it, expect, beforeEach, vi, type MockInstance } from 'vitest';
 import createStateManager from '@8f4e/state-manager';
 
-import compiler from './compiler';
+import compiler from './effect';
 
-import configEffect from '../config-compiler/config';
+import configEffect from '../config-compiler/effect';
 import { compileConfigForExport } from '../config-compiler/compileConfigForExport';
 import { createMockState, createMockCodeBlock } from '../../pureHelpers/testingUtils/testUtils';
 import { createMockEventDispatcherWithVitest } from '../../pureHelpers/testingUtils/vitestTestUtils';
